# Swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "He thong quan ly API cho rap chieu",
    "version": "1.0.0",
    "description": "He thong tai lieu API quan ly rap chieu phim tap trung."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "OK"
          },
          "message": {
            "type": "string",
            "example": "Cinema API is running smoothly!"
          }
        }
      },
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Inception"
          },
          "director": {
            "type": "string",
            "example": "Christopher Nolan"
          }
        }
      },
      "MovieInput": {
        "type": "object",
        "required": [
          "title",
          "director"
        ],
        "properties": {
          "title": {
            "type": "string",
            "example": "Interstellar"
          },
          "director": {
            "type": "string",
            "example": "Christopher Nolan"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Khong tim thay phim."
          }
        }
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Kiem tra tinh trang hoat dong cua server",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "Server hoat dong binh thuong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/movie": {
      "get": {
        "summary": "Lay danh sach phim",
        "tags": [
          "Movies"
        ],
        "responses": {
          "200": {
            "description": "Tra ve danh sach phim",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Them phim moi",
        "tags": [
          "Movies"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieInput"
              },
              "examples": {
                "newMovie": {
                  "summary": "Phim moi",
                  "value": {
                    "title": "Interstellar",
                    "director": "Christopher Nolan"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Da them phim",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          },
          "400": {
            "description": "Du lieu gui len khong hop le",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/movie/{id}": {
      "get": {
        "summary": "Lay thong tin mot phim theo ID",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1,
            "description": "ID cua phim can xem"
          }
        ],
        "responses": {
          "200": {
            "description": "Tra ve thong tin phim",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          },
          "400": {
            "description": "ID phim khong hop le",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Khong tim thay phim",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Cap nhat thong tin phim",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1,
            "description": "ID cua phim can cap nhat"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieInput"
              },
              "examples": {
                "updateMovie": {
                  "summary": "Thong tin cap nhat",
                  "value": {
                    "title": "Inception",
                    "director": "Christopher Nolan"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Da cap nhat phim",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          },
          "400": {
            "description": "ID hoac du lieu gui len khong hop le",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Khong tim thay phim",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Xoa phim theo ID",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1,
            "description": "ID cua phim can xoa"
          }
        ],
        "responses": {
          "204": {
            "description": "Da xoa phim"
          },
          "400": {
            "description": "ID phim khong hop le",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Khong tim thay phim",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}